+++ src/main/java/redstonedubstep/mods/vanishmod/VanishUtil.java
@@ -58,10 +58,7 @@ public class VanishUtil {
 	}
 
 	public static void updateVanishedStatus(ServerPlayerEntity player, boolean vanished) {
-		CompoundNBT deathPersistedData = new CompoundNBT();
-
-		deathPersistedData.putBoolean("Vanished", vanished);
-		player.getPersistentData().put(PlayerEntity.PERSISTED_NBT_TAG, deathPersistedData);
+		player.getPersistentData().getCompound(PlayerEntity.PERSISTED_NBT_TAG).putBoolean("Vanished", vanished);
 		player.setInvisible(vanished);
 
 		if (ModList.get().isLoaded("minecraft2discord")) {